Don't update institution logo when it comes null. Update conclusion data provider to compare conclusion date with a date. Update Language helper to know the locale to use in a document

+++ src/main/java/org/fenixedu/qubdocs/ui/institutionconfiguration/InstitutionConfigurationController.java
@@ -82,6 +82,10 @@ public class InstitutionConfigurationController extends FenixeduQubdocsReportsBa
 
         configuration.setName(institutionName);
 
+        if (logoFile.getBytes().length == 0) {
+            return;
+        }
+
         if (configuration.getInstitutionLogo() != null) {
             configuration.getInstitutionLogo().delete();
         }
